refactor TSPTests
